fix(stats): widen the rate accumulator so fast links stop wrapping it
CPreciseRateCounter::m_total held the sum of every sample in the window
in a uint32. With count_average the samples are rates, not byte counts,
and the graphs' running average keeps a 5 minute window of them -- 100
samples at the default 3 s spacing -- so the sum passes 2^32 once the
mean reaches about 41 MB/s. Past that it wrapped, the eviction loop's
subtraction drove it further under, and the trend collapsed to near
nothing and climbed back over a full averaging window, repeatedly, for
as long as the transfer stayed fast.

Reproduced against the real arithmetic: at 45000 kB/s the average
reports 3.5 MB/s instead of 46, at 56000 kB/s it reports 14.8 instead
of 57; both are exact at 64 bits. On a graph the collapses all begin at
the same height, which is what identified it -- a fixed value, not a
fixed interval.

The statistics tree's Max download/upload rate counters share the class.
They sum actual bytes over 30 s rather than rates over 5 minutes, so
their ceiling was about 143 MB/s; the same widening lifts it.

Only the sum was too narrow -- the individual samples stay 32 bits.

+++ src/Statistics.h
@@ -168,7 +168,15 @@ protected:
 	std::deque<uint32> m_byte_history;
 	std::deque<uint64> m_tick_history;
 	uint32_t m_timespan;
-	uint32_t m_total;
+	// Sum of every sample currently inside the window. Wide because the
+	// sum is not a byte count: with count_average the samples are rates,
+	// and the graphs' running average holds a 5 minute window of them --
+	// 100 samples at the default 3 s spacing, which passes 2^32 once the
+	// mean rate reaches about 41 MB/s. Past that it wrapped, the
+	// subtraction below drove it further under, and the trend collapsed
+	// to nothing and climbed back over a full window. The individual
+	// samples in m_byte_history are fine at 32 bits; only their sum is not.
+	uint64_t m_total;
 	double m_rate;
 	double m_max_rate;
 	uint32_t m_tmp_sum;
